Fixed a segfault. Oooops.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@503 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Text/Text.cpp
@@ -311,5 +311,5 @@ void Text::mouseHighlight(int x, int y) {
 }
 
 void Text::mouseHighlight(double x, double y) {
-   mouseHighlight(x, y);
+   mouseHighlight((int) x, (int) y);
 }
